Refactor camera overlay - PR feedback
