+++ src/components/Theme/style.css
@@ -82,22 +82,6 @@
   text-align: center;
 }
 
-.overlay-shape {
-  border: 1px solid #00c8a0;
-  position: absolute;
-  bottom: 50%;
-  left: 50%;
-  transform: translate(-50%, 50%);
-  /* 300px box-shadow, so 600px by 600px, so as to definitely cover the modal */
-  box-shadow: 0 0 0 300px rgba(0, 0, 0, 0.7);
-}
-
-.overlay {
-  display: flex;
-  width: 100%;
-  height: 100%;
-}
-
 .icon {
   display: block;
   background-repeat: no-repeat;
